Move quoteShellArgs to command
