+++ packages/create-cloudflare/templates/qwik/c3.ts
@@ -3,11 +3,10 @@ import { brandColor } from "@cloudflare/cli/colors";
 import { spinner } from "@cloudflare/cli/interactive";
 import { runFrameworkGenerator } from "frameworks/index";
 import { loadTemplateSnippets, transformFile } from "helpers/codemod";
-import { runCommand } from "helpers/command";
+import { quoteShellArgs, runCommand } from "helpers/command";
 import { usesTypescript } from "helpers/files";
 import { detectPackageManager } from "helpers/packageManagers";
 import * as recast from "recast";
-import { quoteShellArgs } from "../../src/common";
 import type { TemplateConfig } from "../../src/templates";
 import type { C3Context } from "types";
